pyproject.toml
[build-system]
requires = ["poetry-core", "ziggy-pydust==0.23.2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ziggy-pydust-template"
version = "0.1.0"
description = ""
authors = ["Fulcrum Inc. <hello@fulcrum.so>"]
readme = "README.md"
packages = [{ include = "zigffi" }]
include = [
    { path = "src/", format = "sdist" },
    { path = "zigffi/*.so", format = "wheel" },
    "zigffi/*.pyi",
]

[tool.poetry.build]
script = "build.py"

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.1.5"
boto3 = "^1.34.71"
psycopg2 = "^2.9.9"
pendulum = "^3.0.0"
numpy = "^1.26.4"
pandas = "^2.2.1"

[tool.poetry.group.dev.dependencies]
ziggy-pydust = "^0.23.2"
pytest = "^8.0.0"
ruff = "^0.3.0"
black = "^23.7.0"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
select = ["F", "E", "W", "UP", "I001", "I002"]

[tool.pydust]
self_managed = true

# [[tool.pydust.ext_module]]
# name = "zigffi._lib"
# root = "src/fib.zig"
